Refactor: split Evaluator into validator & dynamic diagram
