+++ src/content/components/landing-page/repl/ReplEvaluator.tsx
@@ -1,56 +1,12 @@
-import './ReplEvaluator.css'
 import { useState, type ChangeEvent, type KeyboardEvent } from 'react'
-import { Package, link, Interpreter, Evaluation, WRENatives, type ExecutionResult, interprete, REPL, WRE, fromJSON, getDynamicDiagramData, type DynamicDiagramElement, type DynamicDiagramNode, type DynamicDiagramReference, LIST_MODULE, SET_MODULE, WOLLOK_EXTRA_STACK_TRACE_HEADER, parse, validate, WOLLOK_FILE_EXTENSION, type Problem } from 'wollok-ts'
-import type { ElementDefinition } from 'cytoscape'
-
-type WollokError = {
-  line: number,
-  column?: number,
-  message: string,
-  type?: 'error' | 'info' | 'warning',
-}
-
-const mostrarError = (error: WollokError) => ({
-    row: error.line - 1,
-    column: error.column ?? 0,
-    text: error.message,
-    type: error.type ?? 'error',
-  })
+import { Evaluation, Interpreter, Package, REPL, WOLLOK_FILE_EXTENSION, WRE, WRENatives, fromJSON, interprete, link, parse, validate, type ExecutionResult } from 'wollok-ts'
+import { getDataDiagram, sanitizeStackTrace } from './replDynamicDiagram'
+import './ReplEvaluator.css'
 
 const buildEnvironment = (aPackage: Package) => {
   const environment = link([aPackage], fromJSON(WRE))
   return new Interpreter(Evaluation.build(environment, WRENatives))
 }
-
-// ************************************************************************************************************
-// TODO: migrate from wollok-lsp-ide > reporter to wollok-web-tools
-const convertToHumanReadable = (code: string) => {
-  if (!code) {
-    return ''
-  }
-  const result = code.replace(
-    /[A-Z0-9]+/g,
-    (match) => ' ' + match.toLowerCase()
-  )
-  return (
-    result.charAt(0).toUpperCase() +
-    result.slice(1, result.length)
-  )
-}
-
-const interpolateValidationMessage = (message: string, ...values: string[]) =>
-  message.replace(/{\d*}/g, (match: string) => {
-    const index = match.replace('{', '').replace('}', '') as unknown as number
-    return values[index] || ''
-  })
-
-const getMessage = (message: string, values: string[]): string =>
-  interpolateValidationMessage(convertToHumanReadable(message), ...values)
-
-const reportValidationMessage = (problem: Problem): string =>
-  getMessage(problem.code, problem.values.concat())
-
-// ************************************************************************************************************
 
 const buildInterpreter = () => {
   try {
@@ -60,12 +16,12 @@ const buildInterpreter = () => {
     const interpreter = buildEnvironment(replPackage)
     const problems = validate(interpreter.evaluation.environment)
     // @ts-ignore
-    mostrarErrores(problems.map(problem => mostrarError({ line: problem.sourceMap?.start.line ?? 1, column: problem.sourceMap?.start.column ?? 0, message: reportValidationMessage(problem) ?? 'Unexpected Error', type: problem.level })))
+    showErrors(problems.map(problem => showProblem(problem)))
     return interpreter
   } catch (e) {
     console.info(e)
     // @ts-ignore
-    mostrarErrores([mostrarError({ line: 1, message: e })])
+    showErrors([showError(e)])
     return buildEnvironment(new Package({ name: REPL }))
   }
 }
@@ -75,67 +31,6 @@ let interpreter = buildInterpreter()
 const interpreteLine = (expression: string) => {
   return interprete(interpreter!, expression)
 }
-
-/**************************************************************************************************************************************************************/
-/* Copied from ts-cli => should be migrated to wollok web tools */
-const getDataDiagram = (interpreter: Interpreter, rootFQN?: Package): ElementDefinition[] =>
-  getDynamicDiagramData(interpreter, rootFQN)
-    .map((dynamicDiagramElement: DynamicDiagramElement) =>
-      dynamicDiagramElement.elementType === 'node' ? convertToCytoscapeNode(dynamicDiagramElement as DynamicDiagramNode) : convertToCytoscapeReference(dynamicDiagramElement as DynamicDiagramReference)
-    )
-
-const convertToCytoscapeNode = ({ id, type, label }: DynamicDiagramNode): ElementDefinition => ({
-  data: {
-    id,
-    label,
-    type,
-    fontsize: getFontSize(label),
-  },
-})
-
-const convertToCytoscapeReference = ({ id, label, sourceId, targetId, sourceModule, constant }: DynamicDiagramReference): ElementDefinition => ({
-  data: {
-    id,
-    label: `${label}${constant ? '🔒' : ''}`,
-    source: sourceId,
-    target: targetId,
-    width: sourceModule ? 1 : 1.5,
-    fontsize: getFontSize(label),
-    style: getStyle(sourceModule ?? ''),
-  },
-})
-
-
-const getFontSize = (text: string): string => {
-  const textWidth = text.length
-  if (textWidth > 8) return '7px'
-  if (textWidth > 5) return '8px'
-  return '9px'
-}
-
-const getStyle = (sourceModule: string) =>
-  [LIST_MODULE, SET_MODULE].includes(sourceModule) ? 'dotted' : 'solid'
-
-
-// Copied from utils.ts - wollok-ts-cli - should move to wollok-web-tools
-export const sanitizeStackTrace = (e?: Error): string => {
-  if (e?.message) {
-    const originalMessage = e.message.split('\n')
-    return originalMessage[0] ?? e.message
-  }
-
-  const indexOfTsStack = e?.stack?.indexOf(WOLLOK_EXTRA_STACK_TRACE_HEADER)
-  const fullStack = e?.stack?.slice(0, indexOfTsStack ?? -1) ?? ''
-
-  return fullStack
-    .replaceAll('\t', '  ')
-    .replaceAll('     ', '  ')
-    .replaceAll('    ', '  ')
-    .split('\n')
-    .filter(stackTraceElement => stackTraceElement.trim())
-    .join('\n')
-}
-/**************************************************************************************************************************************************************/
 
 export const ReplEvaluator = () => {
   const [expression, setExpression] = useState('')
